#489 - Refactor work package backend
