+++ src/backend/src/utils/errors.utils.ts
@@ -54,4 +54,4 @@ export const errorHandler: ErrorRequestHandler = (error: unknown, _req: Request,
 };
 
 // type so that the not found error messages are consistent
-type NotFoundObjectNames = 'User' | 'Risk' | 'Work Package' | 'Project' | 'Description Bullet' | 'Team';
+type NotFoundObjectNames = 'User' | 'Risk' | 'Change Request' | 'Work Package' | 'Project' | 'Description Bullet' | 'Team';
